refactor : switch문 변경을 위해 class로 변환
-  SectionPathTypeCondition 인터페이스를 선언하고 사용
