+++ src/main/java/nextstep/subway/domain/sectiontype/SectionPathType.java
@@ -1,0 +1,7 @@
+package nextstep.subway.domain.sectiontype;
+
+public enum SectionPathType {
+    DISTANCE, DURATION;
+
+
+}
