+++ src/test/java/nextstep/subway/unit/SubwayMapTest.java
@@ -1,7 +1,11 @@
 package nextstep.subway.unit;
 
 import com.google.common.collect.Lists;
-import nextstep.subway.domain.*;
+import nextstep.subway.domain.Line;
+import nextstep.subway.domain.Path;
+import nextstep.subway.domain.Station;
+import nextstep.subway.domain.SubwayMap;
+import nextstep.subway.domain.sectiontype.SectionPathType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.util.ReflectionTestUtils;
